add celery to other features
