+++ app/main.py
@@ -1,6 +1,6 @@
-from fastapi import FastAPI
+from app.api import file_router, resume_parser_router, jd_quantifier_router, grader_summarizer_router, task_status_router
 from fastapi.middleware.cors import CORSMiddleware
-from app.api import file_router, resume_parser_router, jd_quantifier_router, grader_summarizer_router
+from fastapi import FastAPI
 
 app = FastAPI()
 
@@ -16,3 +16,4 @@ app.include_router(file_router.router)
 app.include_router(resume_parser_router.router)
 app.include_router(jd_quantifier_router.router)
 app.include_router(grader_summarizer_router.router)
+app.include_router(task_status_router.router)
